Add code to only allow CRUD functionality to admin

+++ products/views.py
@@ -4,6 +4,7 @@ from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -85,9 +86,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """Add products to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Access not permitted.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -105,9 +110,12 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
     """Edit a specify product within the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Access not permitted.')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, id=product_id)
     if request.method == 'POST':
@@ -130,9 +138,12 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
     """Delete a specify product within the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Access not permitted.')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, id=product_id)
     product.delete()
